+++ simplegui/sg_panel.cpp
@@ -115,7 +115,7 @@ void SG_Panel::SetTexturator(ST_Texturator *stt, int st) {
   }
 
 
-void SG_Panel::SetAspectRatio(double asp) {
+void SG_Panel::SetAspectRatio(float asp) {
   aspect_ratio = asp;
   for(int tx = 0; tx < int(texture.size()); ++tx) {
     texture[tx].SetTextAspectRatio(asp);
