+++ simplegui/sg_passthrough.cpp
@@ -286,13 +286,13 @@ void SG_PassThrough::SetMenu(int but, const vector<string> itms) {
     }
   }
 
-void SG_PassThrough::SetAspectRatio(double asp) {
+void SG_PassThrough::SetAspectRatio(float asp) {
   aspect_ratio = asp;
   if(background) background->SetAspectRatio(aspect_ratio);
   if(widgets.size() > 0) {
     CalcGeometry();
     widgets[0]->AdjustGeometry(&cur_geom);
-    double newaspect = aspect_ratio * cur_geom.xs / cur_geom.ys;
+    float newaspect = aspect_ratio * cur_geom.xs / cur_geom.ys;
     widgets[0]->SetAspectRatio(newaspect);
     }
   }
